Visual Exterior Init Create

+++ views/workorder_show.php
@@ -3,13 +3,17 @@
 </div>
 
 <div class="input-line py-1.5">
-    <div class="w-8/12 mr-1">
+    <div class="w-6/12 mr-1">
         <span class="font-bold pr-1">Cliente:</span>
         <?php echo $order_info['customer']; ?>
     </div>
-    <div class="w-3/12 mr-1">
+    <div class="w-2/12 mr-1">
         <span class="font-bold pr-1">Data:</span>
         <?php echo date('d/m/Y H:i', strtotime($order_info['date_os'])); ?>
+    </div>
+    <div class="w-3/12 mr-1">
+        <span class="font-bold pr-1">Status:</span>
+        <?php echo ($order_info['status'] == 1) ? 'Concluído' : 'Pendente'; ?>
     </div>
     <div class="w-1/12 text-right">
         <span class="font-bold pr-1">OS Nº:</span>
@@ -74,21 +78,40 @@
     </div>
 </div>
 
-<div class="input-line p-1.5 flex-col bg-slate-200 rounded-md">
-    <?php if ($order_info['aditional_info'] != '') : ?>
+<?php if ($order_info['aditional_info'] != '') : ?>
+    <div class="input-line flex-col">
         <div class="w-full font-bold">
             Observações Gerais:
         </div>
         <div class="w-full">
             <?php echo $order_info['aditional_info']; ?>
         </div>
-    <?php endif; ?>
+    </div>
+<?php endif; ?>
+
+<div class="input-line p-2 my-2 bg-slate-200 rounded-md">
+    <div class="grid grid-cols-3 gap-4 w-full">
+        <a href="<?php echo BASE_URL; ?>workorder/visualexterior_create/<?php echo $order_info['id']; ?>">
+            <div class="w-full bg-slate-300 rounded-md p-1.5">Visual Exterior</div>
+        </a>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Pneus</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Freios</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Direção</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Combustível</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Motor</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Suspensão</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Mecânica Geral</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Luzes e Vidros</div>
+        <div class="w-full bg-slate-300 rounded-md p-1.5">Visual Interior</div>
+
+    </div>
 </div>
+
 <div class="input-line justify-center items-center py-3">
     <span class="px-1">
-        <a href="<?php echo BASE_URL; ?>workorder/edit/<?php echo $order_info['id']; ?>" class="btn">Editar</a>
+        <a href="<?php echo BASE_URL; ?>workorder/edit/<?php echo $order_info['id']; ?>" class="btn"><i class="fa-solid fa-up-right-from-square"></i> Editar</a>
     </span>
     <span class="px-1">
-        <a href="<?php echo BASE_URL; ?>workorder/destroy/<?php echo $order_info['id']; ?>" class="btn-danger hover:btn-danger--hover">Excluir</a>
+        <a href="<?php echo BASE_URL; ?>workorder/destroy/<?php echo $order_info['id']; ?>" class="btn-danger hover:btn-danger--hover" onclick="return confirm('Deseja realmente excluir?')"><i class="fa-regular fa-trash-can"></i> Excluir</a>
     </span>
 </div>
